+++ lib/fledex/color/functions.ex
@@ -22,7 +22,7 @@ defmodule Fledex.Color.Functions do
   def create_rainbow_circular_hsv(0, _initial_hue, _reversed), do: []
 
   def create_rainbow_circular_hsv(num_leds, initial_hue, reversed) do
-    hue_change = Kernel.trunc(65_535 / num_leds)
+    hue_change = div(65_535, num_leds)
 
     for n <- 0..(num_leds - 1)//1 do
       %HSV{h: initial_hue + step(reversed, (n * hue_change) >>> 8) &&& 0xFF, s: 240, v: 255}
@@ -65,9 +65,9 @@ defmodule Fledex.Color.Functions do
     bdist87 = (eb - sb) <<< 7
 
     steps = num_leds + 1
-    rdelta = trunc(rdist87 / steps) * 2
-    gdelta = trunc(gdist87 / steps) * 2
-    bdelta = trunc(bdist87 / steps) * 2
+    rdelta = div(rdist87, steps) * 2
+    gdelta = div(gdist87, steps) * 2
+    bdelta = div(bdist87, steps) * 2
 
     r88 = sr <<< 8
     g88 = sg <<< 8